FIX docu and code clenapu
