+++ src/lib/ngsi/ContextAttribute.cpp
@@ -42,6 +42,7 @@
 #include "rest/OrionError.h"
 #include "parse/CompoundValueNode.h"
 
+#include "mongoBackend/location.h"
 #include "mongoBackend/dbConstants.h"
 #include "mongoBackend/dbFieldEncoding.h"
 #include "mongoBackend/compoundValueBson.h"
@@ -930,62 +931,6 @@ void ContextAttribute::filterAndOrderMetadata
 }
 
 
-static orion::CompoundValueNode* getGeometryFromFeature(orion::CompoundValueNode* feature)
-{
-  for (unsigned int ix = 0; ix < feature->childV.size(); ++ix)
-  {
-    orion::CompoundValueNode* childP = feature->childV[ix];
-    if (childP->name == "geometry")
-    {
-      return childP;
-    }
-  }
-
-  LM_E(("Runtime Error (geometry field expected in GeoJson Features)"));
-  return NULL;
-}
-
-
-static orion::CompoundValueNode* getGeometryFromFeatureCollection(orion::CompoundValueNode* featureCollection)
-{
-  for (unsigned int ix = 0; ix < featureCollection->childV.size(); ++ix)
-  {
-    orion::CompoundValueNode* childP = featureCollection->childV[ix];
-    if (childP->name == "features")
-    {
-      return getGeometryFromFeature(featureCollection->childV[ix]->childV[0]);
-    }
-  }
-
-  LM_E(("Runtime Error (feature field expected in GeoJson FeatureCollection)"));
-  return NULL;
-}
-
-
-
-static orion::CompoundValueNode* getGeometryToRender(orion::CompoundValueNode* compoundValueP)
-{
-  for (unsigned int ix = 0; ix < compoundValueP->childV.size(); ++ix)
-  {
-     orion::CompoundValueNode* childP = compoundValueP->childV[ix];
-     if ((childP->name == "type") && (childP->valueType == orion::ValueTypeString))
-     {
-       if (childP->stringValue == "Feature")
-       {
-         return getGeometryFromFeature(compoundValueP);
-       }
-       if (childP->stringValue == "FeatureCollection")
-       {
-         return getGeometryFromFeatureCollection(compoundValueP);
-       }
-     }
-  }
-
-  // Regular geo:json
-  return compoundValueP;
-}
-
-
 
 /* ****************************************************************************
 *
@@ -1018,7 +963,7 @@ std::string ContextAttribute::toJson(const std::vector<std::string>&  metadataFi
     orion::CompoundValueNode* childToRenderP = compoundValueP;
     if (type == GEO_JSON)
     {
-      childToRenderP = getGeometryToRender(compoundValueP);
+      childToRenderP = getGeometry(compoundValueP);
     }
 
     // Some internal error conditions in getGeometryToRender() (e.g. out of band manipulation
